+++ Changelog.md
@@ -9,6 +9,33 @@ CHANGELOG
 Updates
 -------
 
+05/04/2020 2.2.0-preview4
+=========================
+
+WinCopies.Util (2.2.0-preview4)
+-------------------------------
+
+- Additions:
+	- Classes:
+		- WinCopies.Linq.Extensions
+
+- Removals:
+	- WinCopies.Util.IDisposable.IsDisposed property, as WinCopies.Util.IDisposable implements WinCopies.Util.DotNetFix.IDisposable, that implements this property.
+
+WinCopies.Util.Desktop (2.2.0-preview4)
+---------------------------------------
+
+- Additions:
+	- Classes:
+		- IconToImageSourceConverter
+	- Extension methods:
+		- ToImageSource(this Bitmap bitmap)
+
+- Existing items behavior updates:
+	- All command-related items have moved to the WinCopies.Util.Commands namespace.
+	- Change the type of the WinCopies.Util.Commands.DelegateCommand.CanExecuteDelegate property for the non-generic class from System.Predicate<object> to WinCopies.Util.Predicate.
+	- Add explicit constructors to the WinCopies.Util.Commands.DelegateCommand classes (generic and non-generic).
+
 04/24/2020 2.2.0-preview3
 =========================